large_file_analysis.py
#!/usr/bin/env python3
"""
대용량 파일 처리 현황 분석 및 테스트
"""
import os
import psutil
import time
from pathlib import Path

def analyze_large_file_processing():
    """대용량 파일 처리 시 예상되는 문제점 분석"""
    
    print("=== 대용량 파일 처리 현황 분석 ===")
    print()
    
    # 현재 설정 확인
    current_limits = {
        "max_file_size_mb": 10,  # 현재 10MB 제한
        "chunk_size": 1000,      # 텍스트 청킹 크기
        "chunk_overlap": 200,    # 청크 오버랩
        "max_concurrent": 8      # 동시 처리 수
    }
    
    print("📋 현재 시스템 제한:")
    for key, value in current_limits.items():
        print(f"  - {key}: {value}")
    print()
    
    # 대용량 파일 시나리오
    large_file_scenarios = [
        {"pages": 50, "size_mb": 15, "content_chars": 150000, "desc": "중간 규모 보고서"},
        {"pages": 100, "size_mb": 30, "content_chars": 300000, "desc": "대규모 매뉴얼"},
        {"pages": 200, "size_mb": 50, "content_chars": 600000, "desc": "종합 기술문서"},
        {"pages": 500, "size_mb": 100, "content_chars": 1500000, "desc": "대용량 규격서"}
    ]
    
    print("📊 대용량 파일 시나리오별 예상 문제:")
    print()
    
    for scenario in large_file_scenarios:
        print(f"🔸 {scenario['desc']} ({scenario['pages']}페이지, {scenario['size_mb']}MB)")
        
        # 문제점 분석
        problems = analyze_scenario_problems(scenario, current_limits)
        
        for problem in problems:
            print(f"  ❌ {problem}")
        
        # 예상 처리 시간
        estimated_time = estimate_processing_time(scenario)
        print(f"  ⏱️  예상 처리 시간: {estimated_time:.1f}초")
        
        # 메모리 사용량 예상
        estimated_memory = estimate_memory_usage(scenario)
        print(f"  💾 예상 메모리 사용량: {estimated_memory:.1f}MB")
        print()

def analyze_scenario_problems(scenario, current_limits):
    """시나리오별 문제점 분석"""
    problems = []
    
    # 1. 파일 크기 제한
    if scenario["size_mb"] > current_limits["max_file_size_mb"]:
        problems.append(f"파일 크기 초과 ({scenario['size_mb']}MB > {current_limits['max_file_size_mb']}MB)")
    
    # 2. 메모리 사용량
    estimated_memory = scenario["content_chars"] * 4 / (1024 * 1024)  # 대략적인 메모리 사용량
    if estimated_memory > 100:  # 100MB 이상
        problems.append(f"높은 메모리 사용량 ({estimated_memory:.1f}MB)")
    
    # 3. 처리 시간
    estimated_time = scenario["content_chars"] / 10000  # 대략적인 처리 시간 (초)
    if estimated_time > 60:  # 1분 이상
        problems.append(f"긴 처리 시간 ({estimated_time:.1f}초)")
    
    # 4. 청킹 수
    chunk_count = scenario["content_chars"] // current_limits["chunk_size"]
    if chunk_count > 1000:  # 1000개 이상 청크
        problems.append(f"과도한 청크 수 ({chunk_count}개)")
    
    # 5. HTTP 타임아웃
    if estimated_time > 300:  # 5분 이상
        problems.append("HTTP 요청 타임아웃 위험")
    
    return problems

def estimate_processing_time(scenario):
    """처리 시간 추정"""
    base_time = 2  # 기본 2초
    content_factor = scenario["content_chars"] / 50000  # 50k 문자당 1초 추가
    nlp_factor = scenario["content_chars"] / 100000  # NLP 처리 시간
    embedding_factor = scenario["content_chars"] / 200000  # 임베딩 생성 시간
    
    return base_time + content_factor + nlp_factor + embedding_factor

def estimate_memory_usage(scenario):
    """메모리 사용량 추정"""
    # 텍스트 메모리 (UTF-8, 4바이트/문자 가정)
    text_memory = scenario["content_chars"] * 4 / (1024 * 1024)
    
    # 임베딩 메모리 (768차원 float32)
    chunk_count = scenario["content_chars"] // 1000
    embedding_memory = chunk_count * 768 * 4 / (1024 * 1024)
    
    # 처리 과정 중 임시 메모리
    temp_memory = text_memory * 2
    
    return text_memory + embedding_memory + temp_memory

def check_system_resources():
    """현재 시스템 리소스 확인"""
    print("🖥️  현재 시스템 리소스:")
    
    # CPU 정보
    cpu_count = psutil.cpu_count()
    cpu_percent = psutil.cpu_percent(interval=1)
    print(f"  - CPU: {cpu_count}코어, 사용률 {cpu_percent:.1f}%")
    
    # 메모리 정보
    memory = psutil.virtual_memory()
    print(f"  - 메모리: {memory.total/(1024**3):.1f}GB 총량, {memory.available/(1024**3):.1f}GB 사용가능")
    
    # 디스크 정보
    disk = psutil.disk_usage('/')
    print(f"  - 디스크: {disk.total/(1024**3):.1f}GB 총량, {disk.free/(1024**3):.1f}GB 사용가능")
    print()

def get_recommendations():
    """대용량 파일 처리 개선 권장사항"""
    print("🚀 대용량 파일 처리 개선 권장사항:")
    print()
    
    recommendations = [
        {
            "category": "즉시 적용 가능",
            "items": [
                "파일 크기 제한 10MB → 100MB로 증가",
                "청킹 전략 개선 (큰 파일은 더 큰 청크)",
                "스트리밍 처리 방식 도입",
                "메모리 사용량 모니터링 강화"
            ]
        },
        {
            "category": "단기 개선 (1-2주)",
            "items": [
                "비동기 파일 업로드 (청크 단위)",
                "백그라운드 처리 (FastAPI BackgroundTasks)",
                "진행률 실시간 추적",
                "파일 압축 및 최적화"
            ]
        },
        {
            "category": "중기 개선 (1-2개월)",
            "items": [
                "Celery 기반 비동기 처리",
                "Redis 기반 작업 큐",
                "분산 처리 아키텍처",
                "S3 기반 파일 스토리지"
            ]
        }
    ]
    
    for rec in recommendations:
        print(f"📂 {rec['category']}:")
        for item in rec['items']:
            print(f"  ✅ {item}")
        print()

if __name__ == "__main__":
    analyze_large_file_processing()
    check_system_resources()
    get_recommendations()
